refactor: refactoring into productinfo molecule
